manage project structure
