+++ accounts/admin.py
@@ -1,0 +1,7 @@
+from django.contrib import admin
+
+from accounts.models import User
+
+# Register your models here.
+admin.site.register(User)
+
