+++ accounts/forms/__init__.py
@@ -1,0 +1,2 @@
+from .login import LoginForm
+from .register import RegisterForm
