+++ accounts/forms/register.py
@@ -1,11 +1,13 @@
 from django import forms
-from django.contrib.auth import authenticate, get_user_model
+from django.contrib.auth import get_user_model
 from django.contrib.auth.forms import UserCreationForm
 from django.forms.renderers import TemplatesSetting
 
-from task_management.constants import HELP_TEXT, MESSAGES
+from accounts.constants import HELP_TEXT, MESSAGES
 
 User = get_user_model()
+
+FORM_TEMPLATE = "accounts/form.html"
 
 
 class RegisterForm(UserCreationForm):
@@ -22,7 +24,7 @@ class RegisterForm(UserCreationForm):
             "password2",
         )
 
-    template_name = "task_management/forms/auth-form.html"
+    template_name = FORM_TEMPLATE
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
@@ -53,44 +55,5 @@ class RegisterForm(UserCreationForm):
                 raise forms.ValidationError(
                     MESSAGES["password_not_match"], code="invalid"
                 )
-
-        return cleaned_data
-
-
-class LoginForm(forms.Form):
-    renderer = TemplatesSetting()
-
-    username = forms.CharField(
-        required=True,
-        widget=forms.TextInput(),
-    )
-
-    password = forms.CharField(
-        required=True,
-        widget=forms.PasswordInput(),
-    )
-
-    template_name = "task_management/forms/auth-form.html"
-
-    def __init__(self, *args, **kwargs):
-        self.user = None
-        super().__init__(*args, **kwargs)
-
-    def clean(self):
-        cleaned_data = super().clean()
-
-        if cleaned_data is not None:
-            username = cleaned_data.get("username")
-            password = cleaned_data.get("password")
-
-            if username and password:
-                user = authenticate(username=username, password=password)
-
-                if user is None:
-                    raise forms.ValidationError(
-                        MESSAGES["invalid_username_or_password"], code="invalid"
-                    )
-
-                self.user = user
 
         return cleaned_data
