+++ accounts/templates/accounts/register.html
@@ -1,4 +1,4 @@
-{% extends 'task_management/auth-layout.html' %}
+{% extends 'accounts/layout.html' %}
 
 {% block title %}Register{% endblock %}
