+++ static/css/base.css
@@ -1,0 +1,2 @@
+@import url("_variables.css");
+@import url("_reset.css");
