--- task_management/forms/__init__.py
@@ -1,1 +1,0 @@
-from .auth_form import RegisterForm, LoginForm
