+++ task_management/models/__init__.py
@@ -4,4 +4,3 @@ from .status import Status
 from .task import Task
 from .task_assignment import TaskAssignment
 from .task_status import TaskStatus
-from .user import User
