+++ task_management/models/project.py
@@ -1,6 +1,6 @@
 from django.db import models
 
-from .user import User
+from accounts.models import User
 
 
 class Project(models.Model):
