--- task_management/static/task_management/css/_auth-layout.css
@@ -1,25 +1,0 @@
-body {
-  display: flex;
-  flex-direction: column;
-  min-height: 100vh;
-}
-
-.header {
-  height: 76px;
-  background-color: var(--white-02);
-  padding:  8px 16px;
-}
-
-.header .logo-icon {
-  width: auto;
-  height: 100%;
-}
-
-.auth-content {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 100%;
-  height: 100%;
-  padding: 0 16px;
-}
