+++ task_management/static/task_management/css/style.css
@@ -1,0 +1,1 @@
+@import url("./_layout.css");
