--- task_management/static/task_management/css/styles.css
@@ -1,3 +1,0 @@
-@import url("./_variables.css");
-@import url("./_reset.css");
-@import url("./_layout.css");
