+++ task_management/urls.py
@@ -4,7 +4,4 @@ from . import views
 
 urlpatterns = [
     path('', views.index, name='index'),
-    path('login', views.login_view, name='login'),
-    path('logout', views.logout_view, name='logout'),
-    path('register', views.register, name='register'),
 ]
